% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project3D.R
\name{Project3D}
\alias{Project3D}
\title{Projects 3D graphing elements to a 2D canvas}
\description{
This is an R6 class that allows establishment of a 3D coordinate
system on a 2D canvas. After the class is instantiated and initiated,
simple base R graphics tools can be used to render projections of 3D
graphs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lims}}{Graphing region limits for each axis in 3D space.
Expected data type is a three-element numeric vector, where the
elements represent the limits in the x, y, and z
directions, respectively.}

\item{\code{phi}}{The altitudinal rotation angle for placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{theta}}{The azimuthal rotation angle for placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{r}}{The radial distance from the centroid of the graph placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{d}}{Parameter controlling the magnitude of the vanishing effect in the perspective of the projection. Values greater than one diminish the perspective effect and values less than one exaggerate the perspective effect. Expected data type is a single-element numeric vector.}

\item{\code{expand}}{Expansion factors for each axis in 3D space.
Expected data type is a three-element numeric vector, where the
elements represent the expansion vectors in the x, y, and z
directions, respectively.}

\item{\code{center}}{Coordinates of the centroid of the graphing volume in 3D space.
Expected data type is a three-element numeric vector, where the
elements represent the coordinates of the centroid in the x, y, and z
directions, respectively.}

\item{\code{size}}{The size of the graphing volume in 3D space.
Expected data type is a three-element numeric vector, where the
elements represent the lengths of the volume in the x, y, and z
directions, respectively.}

\item{\code{vt}}{The visual transformation matrix.
Expected data type is a four by four numeric matrix.}

\item{\code{vtt}}{The transpose of the visual transformation matrix
Expected data type is a four by four numeric matrix.}

\item{\code{lims_2d}}{The limits of the graphing volume projected to 2D space.
Expected data type is a two-element list of two-element numeric
vectors. The elements of the list correspond to the limits in the
x and y directions, respectively. The elements of each vector in the
list correspond to the minimum and maximum values for a given axis,
respectively.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Project3D-new}{\code{Project3D$new()}}
\item \href{#method-Project3D-createIdentity}{\code{Project3D$createIdentity()}}
\item \href{#method-Project3D-transform}{\code{Project3D$transform()}}
\item \href{#method-Project3D-translate}{\code{Project3D$translate()}}
\item \href{#method-Project3D-scale}{\code{Project3D$scale()}}
\item \href{#method-Project3D-rotate.x}{\code{Project3D$rotate.x()}}
\item \href{#method-Project3D-rotate.y}{\code{Project3D$rotate.y()}}
\item \href{#method-Project3D-perspective}{\code{Project3D$perspective()}}
\item \href{#method-Project3D-mode.default}{\code{Project3D$mode.default()}}
\item \href{#method-Project3D-window.xy}{\code{Project3D$window.xy()}}
\item \href{#method-Project3D-coords.xy}{\code{Project3D$coords.xy()}}
\item \href{#method-Project3D-box.x}{\code{Project3D$box.x()}}
\item \href{#method-Project3D-box.y}{\code{Project3D$box.y()}}
\item \href{#method-Project3D-box.z}{\code{Project3D$box.z()}}
\item \href{#method-Project3D-text}{\code{Project3D$text()}}
\item \href{#method-Project3D-ticks.x}{\code{Project3D$ticks.x()}}
\item \href{#method-Project3D-ticks.y}{\code{Project3D$ticks.y()}}
\item \href{#method-Project3D-ticks.z}{\code{Project3D$ticks.z()}}
\item \href{#method-Project3D-clone}{\code{Project3D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new object that is an instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$new(
  xlim,
  ylim,
  zlim,
  phi,
  theta,
  r,
  d = 1,
  xexp = 1,
  yexp = 1,
  zexp = 1,
  scale = TRUE,
  mode = "default"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{Limits of the graphing volume in the x direction.
Expected data type is a two-element numeric vector with the
first element as the minimum and the second element as the
maximum.}

\item{\code{ylim}}{Limits of the graphing volume in the y direction
Expected data type is a two-element numeric vector with the
first element as the minimum and the second element as the
maximum.}

\item{\code{zlim}}{Limits of the graphing volume in the z direction
Expected data type is a two-element numeric vector with the
first element as the minimum and the second element as the
maximum.}

\item{\code{phi}}{The altitudinal rotation angle for placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{theta}}{The azimuthal rotation angle for placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{r}}{The radial distance from the centroid of the graph placing the observer in 3D space relative to the centroid of the graphing volume. Expected data type is a single-element numeric vector.}

\item{\code{d}}{Parameter controlling the magnitude of the vanishing effect in the perspective of the projection. Values greater than one diminish the perspective effect and values less than one exaggerate the perspective effect. Expected data type is a single-element numeric vector.

Default value is 1.}

\item{\code{xexp}}{Expansion factor in the x direction.
Allows adjustment of the relationship of the rendered length
of the axis to the limits of the scale. A value less than one will
shorten the axis relative to the default scale and a value greater
than one will lengthen the axis relative to the default scale.
Expected type is a single-element numeric vector.

Default value is 1.}

\item{\code{yexp}}{Expansion factor in the y direction.
Allows adjustment of the relationship of the rendered length
of the axis to the limits of the scale. A value less than one will
shorten the axis relative to the default scale and a value greater
than one will lengthen the axis relative to the default scale.
Expected type is a single-element numeric vector.

Default value is 1.}

\item{\code{zexp}}{Expansion factor in the z direction.
Allows adjustment of the relationship of the rendered length
of the axis to the limits of the scale. A value less than one will
shorten the axis relative to the default scale and a value greater
than one will lengthen the axis relative to the default scale.
Expected type is a single-element numeric vector.

Default value is 1.}

\item{\code{scale}}{Flag indicating if the axes should maintain default lengths
independent of their scales.
A value of TRUE indicates axes of common size
independent of scale, and a value of FALSE indicates
axes with differing sizes dependent on scales.
Note that the expansion factor arguments xexp, yexp, and zexp
provide an independent method of controlling the size of an axis
relative to its scale.
Expected data type is a single-element logical vector.

Default value is TRUE.}

\item{\code{mode}}{Name of the mode for calculating the visual transformation matrix.
Expected data type is a single-element character string vector.

Default value is "default".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A reference to the new object as an R6 extension of an environment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-createIdentity"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-createIdentity}{}}}
\subsection{Method \code{createIdentity()}}{
Creates an identity matrix with all zeros except for ones on the
diagonal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$createIdentity()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A four by four identity matrix.
Data type is a four-by-four element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-transform}{}}}
\subsection{Method \code{transform()}}{
Transforms the current visual transformation matrix according to the
provided matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$transform(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{A four by four matrix used in the transformation.
Expected data type is a four-by-four-element numeric matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-translate"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-translate}{}}}
\subsection{Method \code{translate()}}{
Transforms the current visual transformation matrix based on a
translation in 3D space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$translate(coords)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{The change in coordinates in 3D space.
Expected data type is a three-element numeric vector, where the
elements correspond to translation in the x, y, and z directions,
respectively.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-scale"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-scale}{}}}
\subsection{Method \code{scale()}}{
Transforms the current visual transformation matrix based on a
rescaling of the 3D axes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$scale(coords)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{The scaling factors for the axes in 3D space.
Expected data type is a three-element numeric vector, where the
elements correspond to scaling factors for the x, y, and z axes,
respectively.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-rotate.x"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-rotate.x}{}}}
\subsection{Method \code{rotate.x()}}{
Transforms the current visual transformation matrix based on a
rotation around the x axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$rotate.x(angle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{angle}}{The angle for rotating around the x axis.
Expected data type is a single-element numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-rotate.y"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-rotate.y}{}}}
\subsection{Method \code{rotate.y()}}{
Transforms the current visual transformation matrix based on a
rotation around the y axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$rotate.y(angle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{angle}}{The angle for rotating around the y axis.
Expected data type is a single-element numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-perspective"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-perspective}{}}}
\subsection{Method \code{perspective()}}{
Transforms the current visual transformation matrix based on a
changing the vanishing point perspective.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$perspective(d)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parameter controlling the magnitude of the vanishing effect in the perspective of the projection. Values greater than one diminish the perspective effect and values less than one exaggerate the perspective effect. Expected data type is a single-element numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-mode.default"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-mode.default}{}}}
\subsection{Method \code{mode.default()}}{
Creates the visual transformation matrix based on the default
projection strategy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$mode.default()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The new visual transformation matrix.
Data type is a four-by-four-element numeric matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-window.xy"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-window.xy}{}}}
\subsection{Method \code{window.xy()}}{
Uses base R graphing function plot.window() to create a 2D
plotting area appropriate for projections of the 3D graphing volume.
Optionally allows rendering of the bounding boxes defining the 3D
graphing volume.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$window.xy(
  xlim = self$lims_2d$x,
  ylim = self$lims_2d$y,
  boxes = FALSE,
  box.xvals = self$lims$x,
  box.xargs = vector(mode = "list"),
  box.yvals = self$lims$y,
  box.yargs = box.xargs,
  box.zvals = self$lims$z,
  box.zargs = box.xargs,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xlim}}{The 2D graphing limits passed to plot.window().

Defaults to the field value for lims_2d$x}

\item{\code{ylim}}{The 2D graphing limits passed to plot.window().

Defaults to the field value for lims_2d$y}

\item{\code{boxes}}{Flag indicating if bounding boxes should be rendered.

Default value is FALSE, which prevents rendering of any bounding
boxes.}

\item{\code{box.xvals}}{Values on the 3D x axis scale where the bounding
boxes should be rendered.
Expected data type is a numerical vector with one or more values.

Default values are the minimum and maximum defined in field lims$x.}

\item{\code{box.xargs}}{List of additional arguments to send to the box.x() method.
Expected data type is a named list where the names represent arguments
passed to the box.x() method.

Default value is an empty list, resulting in no additional arguments
being passed to box.x().}

\item{\code{box.yvals}}{Values on the 3D y axis scale where the bounding
boxes should be rendered.
Expected data type is a numerical vector with one or more values.

Default values are the minimum and maximum defined in field lims$y.}

\item{\code{box.yargs}}{List of additional arguments to send to the box.y() method.
Expected data type is a named list where the names represent arguments
passed to the box.y() method.

Default value is an empty list, resulting in no additional arguments
being passed to box.y().}

\item{\code{box.zvals}}{Values on the 3D z axis scale where the bounding
boxes should be rendered.
Expected data type is a numerical vector with one or more values.

Default values are the minimum and maximum defined in field lims$z.}

\item{\code{box.zargs}}{List of additional arguments to send to the box.z() method.
Expected data type is a named list where the names represent arguments
passed to the box.z() method.

Default value is an empty list, resulting in no additional arguments
being passed to box.z().}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics
plot.window() function used to establish the projected
2D coordinate system}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-coords.xy"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-coords.xy}{}}}
\subsection{Method \code{coords.xy()}}{
Calculate the sets of projected 2D coordinates corresponding to
the provided sets of 3D coordinates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$coords.xy(x, y, z)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{3D x coordinates to be projected.
Expected data type is a numeric vector matching the length of the
y and z arguments.}

\item{\code{y}}{3D y coordinates to be projected.
Expected data type is a numeric vector matching the length of the
x and z arguments.}

\item{\code{z}}{3D z coordinates to be projected.
Expected data type is a numeric vector matching the length of the
x and y arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the projected 2D x and y coordinates.
Data type is a two-element list, each of which is a numerical vector
of values for the 2D x and y axes, respectively. The lenght of the
vectors will match the length of the provided 3D coordinate vectors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-box.x"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-box.x}{}}}
\subsection{Method \code{box.x()}}{
Draws boxes normal to the 3D x axis extending to the y and z limits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$box.x(xvals, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xvals}}{Values on the x axis at which boxes should be drawn.
Expected data type is a numeric vector.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics lines()
function used to render the boxes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-box.y"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-box.y}{}}}
\subsection{Method \code{box.y()}}{
Draws boxes normal to the 3D y axis extending to the x and z limits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$box.y(yvals, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yvals}}{Values on the y axis at which boxes should be drawn.
Expected data type is a numeric vector.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics lines()
function used to render the boxes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-box.z"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-box.z}{}}}
\subsection{Method \code{box.z()}}{
Draws boxes normal to the 3D z axis extending to the x and y limits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$box.z(zvals, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zvals}}{Values on the z axis at which boxes should be drawn.
Expected data type is a numeric vector.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics lines()
function used to render the boxes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-text"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-text}{}}}
\subsection{Method \code{text()}}{
Adds text to a graph at locations defined in 3D space.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$text(xvals, yvals, zvals, labels, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xvals}}{Coordinates for text on the x axis in 3D space.
Expected data type is a numeric vector of the same length as the
y and z arguments.}

\item{\code{yvals}}{Coordinates for text on the y axis in 3D space.
Expected data type is a numeric vector of the same length as the
x and z arguments.}

\item{\code{zvals}}{Coordinates for text on the z axis in 3D space.
Expected data type is a numeric vector of the same length as the
x and y arguments.}

\item{\code{labels}}{Vector of text elements to be rendered on the graph.
Expected data type is a character string vector of the same length as
the x, y, and z arguments.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics text()
function used to render the text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-ticks.x"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-ticks.x}{}}}
\subsection{Method \code{ticks.x()}}{
Draw tick mars along a line defined in the x direction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$ticks.x(at, ylims, zlims, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Locations on the x axis for the tick marks.
Expected data type is a numeric vector.}

\item{\code{ylims}}{A vector of the minimum and maximum extent of the tick marks in the
y direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{zlims}}{A vector of the minimum and maximum extent of the tick marks in the
z direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics
segments() functions used to render the tick marks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-ticks.y"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-ticks.y}{}}}
\subsection{Method \code{ticks.y()}}{
Draw tick mars along a line defined in the y direction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$ticks.y(at, xlims, zlims, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Locations on the y axis for the tick marks.
Expected data type is a numeric vector.}

\item{\code{xlims}}{A vector of the minimum and maximum extent of the tick marks in the
x direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{zlims}}{A vector of the minimum and maximum extent of the tick marks in the
z direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics
segments() functions used to render the tick marks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-ticks.z"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-ticks.z}{}}}
\subsection{Method \code{ticks.z()}}{
Draw tick mars along a line defined in the z direction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$ticks.z(at, xlims, ylims, xpd = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{Locations on the z axis for the tick marks.
Expected data type is a numeric vector.}

\item{\code{xlims}}{A vector of the minimum and maximum extent of the tick marks in the
x direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{ylims}}{A vector of the minimum and maximum extent of the tick marks in the
y direction.
Expected data type is a two-element numeric vector, where the first
element is the minimum value and the second element is the
maximum value.}

\item{\code{xpd}}{Value for the base R graphics parameter xpd.Expected value is a single-element logical vector.

Default value is TRUE to be sure the rendered element will not be croppedto the 2D graphing area by the graphics device.}

\item{\code{...}}{Generic parameters that will be passed to the base R graphics
segments() functions used to render the tick marks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a NULL reference.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project3D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Project3D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project3D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
